Dylan/s3en 2127 loading states for all pages (#57)
* loading for data set

* adding loading page for promptset

* dataset loading

* prompts loading

* prompts loading

* metrics loading, chart loading

* evaluations loading

* refactor traces

* trace row loading

* more loading

* Rename loading to skeleton

* Minor fix

---------
