+++ app/(protected)/project/[project_id]/datasets/dataset/[dataset_id]/page.tsx
@@ -2,6 +2,7 @@
 
 import { CreateData } from "@/components/project/dataset/create-data";
 import { EditData } from "@/components/project/dataset/edit-data";
+import DatasetRowSkeleton from "@/components/project/dataset/dataset-row-skeleton";
 import { Spinner } from "@/components/shared/spinner";
 import { Button } from "@/components/ui/button";
 import { Separator } from "@/components/ui/separator";
@@ -69,7 +70,7 @@ export default function Dataset() {
   });
 
   if (fetchDataset.isLoading || !fetchDataset.data || !currentData) {
-    return <div>Loading...</div>;
+    return <PageSkeleton />;
   } else {
     return (
       <div className="w-full py-6 px-6 flex flex-col gap-4">
@@ -129,4 +130,33 @@ export default function Dataset() {
       </div>
     );
   }
+}
+
+function PageSkeleton() {
+  return (
+    <div className="w-full py-6 px-6 flex flex-col gap-4">
+      <div className="flex gap-4 items-center w-fit">
+        <Button
+          disabled={true}
+          variant="secondary"
+          onClick={() => window.history.back()}
+        >
+          <ChevronLeft className="mr-1" />
+          Back
+        </Button>
+        <CreateData disabled={true} />
+      </div>
+      <div className="flex flex-col gap-3 rounded-md border border-muted max-h-screen overflow-y-scroll">
+        <div className="grid grid-cols-5 items-center justify-stretch gap-3 py-3 px-4 bg-muted">
+          <p className="text-xs font-medium">Created at</p>
+          <p className="text-xs font-medium">Input</p>
+          <p className="text-xs font-medium">Output</p>
+          <p className="text-xs font-medium text-end">Note</p>
+        </div>
+        {Array.from({ length: 5 }).map((_, index) => (
+          <DatasetRowSkeleton key={index} />
+        ))}
+      </div>
+    </div>
+  );
 }
