+++ components/project/metrics.tsx
@@ -3,8 +3,10 @@
 import { useParams } from "next/navigation";
 import { useQuery } from "react-query";
 import { EvalChart } from "../charts/eval-chart";
+import LargeChartSkeleton from "../charts/large-chart-skeleton";
 import { TraceLatencyChart } from "../charts/latency-chart";
 import { ModelAccuracyChart } from "../charts/model-accuracy-chart";
+import SmallChartSkeleton from "../charts/small-chart-skeleton";
 import { CostChart, TokenChart } from "../charts/token-chart";
 import { TraceSpanChart } from "../charts/trace-chart";
 import { Info } from "../shared/info";
@@ -37,7 +39,7 @@ export default function Metrics({ email }: { email: string }) {
     fetchTests.isLoading ||
     !fetchTests.data
   ) {
-    return <div>Loading...</div>;
+    return <PageSkeleton />;
   } else {
     // get test obj of factual accuracy test
     const test = fetchTests?.data?.tests?.find(
@@ -81,4 +83,44 @@ export default function Metrics({ email }: { email: string }) {
       </div>
     );
   }
+}
+
+function PageSkeleton() {
+  return (
+    <div className="w-full flex flex-col gap-6 p-6">
+      <div className="flex flex-col gap-2">
+        <p className="text-lg font-semibold">Usage</p>
+        <Separator />
+        <div className="flex flex-row items-center gap-5">
+          <SmallChartSkeleton />
+          <SmallChartSkeleton />
+          <SmallChartSkeleton />
+        </div>
+      </div>
+      <div className="flex flex-row gap-4 w-full">
+        <div className="flex flex-col gap-2 w-full">
+          <div className="flex flex-row items-center gap-2">
+            <p className="text-lg font-semibold">Latency</p>
+          </div>
+          <Separator />
+          <LargeChartSkeleton />
+        </div>
+      </div>
+      <div className="flex flex-col gap-2 w-full">
+        <div className="flex flex-row items-center gap-2">
+          <p className="text-lg font-semibold">Evaluated Accuracy</p>
+          <Info information="This is calculated based on your evaluation of the q&a pairs. Go to the Eval tab to start evaluating to see this metric calculated." />
+        </div>
+        <Separator />
+        <LargeChartSkeleton />
+      </div>
+      <div className="flex flex-col gap-2">
+        <p className="text-lg font-semibold">Evaluated Accuracy per Model</p>
+        <Separator />
+        <div className="flex flex-row items-center gap-5 w-full">
+          <LargeChartSkeleton />
+        </div>
+      </div>
+    </div>
+  );
 }
