+++ components/project/playground.tsx
@@ -3,11 +3,10 @@
 import { ApiKeyDialog } from "@/components/apiKey/api-dialog";
 import { useParams } from "next/navigation";
 import { useQuery } from "react-query";
+import { Skeleton } from "../ui/skeleton";
 
 export default function Playground({ email }: { email: string }) {
   const project_id = useParams()?.project_id as string;
-
-
 
   const fetchProject = useQuery({
     queryKey: ["fetch-project-query"],
@@ -33,7 +32,7 @@ export default function Playground({ email }: { email: string }) {
     fetchUser.isLoading ||
     !fetchUser.data
   ) {
-    return <div>Loading...</div>;
+    return <PageLoading />;
   }
 
   return (
@@ -45,6 +44,19 @@ export default function Playground({ email }: { email: string }) {
         <div>
           <ApiKeyDialog project_id={project_id} />
         </div>
+      </div>
+    </div>
+  );
+}
+
+function PageLoading() {
+  return (
+    <div className="w-full h-full flex items-center justify-center">
+      <div className="flex flex-col items-center gap-4">
+        <div className="w-16 h-16">
+          <Skeleton className="w-full h-full" />
+        </div>
+        <Skeleton className="w-32 h-6" />
       </div>
     </div>
   );
